remove several actual duplicate fns
